fix: put each export in its own file and only export what is actually needed
