+++ src/modules/orgUnit/getOuLevelId.js
@@ -1,0 +1,3 @@
+import { LEVEL_ID_PREFIX } from './internal/constants'
+
+export const getOuLevelId = id => `${LEVEL_ID_PREFIX}-${id}`
