+++ src/modules/orgUnit/internal/constants.js
@@ -1,0 +1,11 @@
+/**
+ * Org unit level id prefix
+ * @type {string}
+ */
+export const LEVEL_ID_PREFIX = 'LEVEL'
+
+/**
+ * Org unit group id prefix
+ * @type {string}
+ */
+export const GROUP_ID_PREFIX = 'OU_GROUP'
